WIP. made progress upto Employer Co-investment Percentage.

+++ src/ESFA.DC.PeriodEnd.ReportService.Model/PeriodEnd/AppsCoInvestment/PaymentInfo.cs
@@ -26,7 +26,7 @@ namespace ESFA.DC.PeriodEnd.ReportService.Model.PeriodEnd.AppsCoInvestment
 
         public short AcademicYear { get; set; }
 
-        public byte TransactionType { get; set; }
+        public int TransactionType { get; set; }
 
         public byte FundingSource { get; set; }
